+++ src/main/java/com/hearthsim/card/spellcard/concrete/WildGrowth.java
@@ -66,7 +66,7 @@ public class WildGrowth extends SpellCard {
 			if (toRet.data_.getCurrentPlayer().getMaxMana() >= 10) {
 				toRet.data_.placeCardHandCurrentPlayer(new ExcessMana());
 			} else {
-				toRet.data_.getCurrentPlayer().addMaxMana(1);
+				toRet.data_.getCurrentPlayer().addMaxMana((byte)1);
 			}
 		}
 		return toRet;
